fix: compability for accessing status enums for users
