+++ modelon/impact/client/entities/__init__.py
@@ -1,0 +1,5 @@
+from modelon.impact.client.entities.status import (
+    CaseStatus,
+    ExperimentStatus,
+    ModelExecutableStatus,
+)
