+++ setup.cfg
@@ -2,3 +2,5 @@
 per-file-ignores =
     modelon/impact/client/__init__.py:F401
     modelon/impact/client/jupyterhub/__init__.py:F401
+    modelon/impact/client/operations/__init__.py:F401
+    modelon/impact/client/entities/__init__.py:F401
